+++ python/masarutils/ui_dbmanager.ui
