+++ test/python/utest_masarclient/test_masarconfigmanager-mongo.py
@@ -1,6 +1,6 @@
 import unittest
 import sqlite3
-from utils.masarconfigmanager import *
+from masarutils.masarconfigmanager import *
 from pymasarsqlite.service.serviceconfig import (saveServiceConfig, retrieveServiceConfigs, retrieveServiceConfigPVs, saveServicePvGroup)
 from masarclient.channelRPC import epicsExit
 from pymasarmongo.db import utils
