+++ test/python/utest_pymasar/test_addmasarconfigs.py
@@ -1,4 +1,4 @@
-from utils.addmasarconfigs import saveMongoService, saveSQLiteServiceConfig, savePvGroups
+from masarutils.addmasarconfigs import saveMongoService, saveSQLiteServiceConfig, savePvGroups
 from pymasarmongo.db import utils
 from pymasarmongo.config._config import masarconfig
 from pymasarsqlite.service.serviceconfig import (retrieveServiceConfigs)
@@ -23,7 +23,7 @@ class TestAddMasarConfigs(unittest.TestCase):
 
     def setUp(self):
         os.chdir(os.path.dirname(os.path.abspath(__file__)))
-        os.chdir("../../../python/utils")
+        os.chdir("../../../python/masarutils")
         self.fdir = os.getcwd()
         json_file = "db_config.txt"
         with open(json_file) as file:
